Added cookie functions in utils
