+++ components/Dark-Theme/useDarkMode.js
@@ -1,26 +1,11 @@
 import { useEffect, useState } from 'react';
+import { setCookie, getCookie } from '../../utils/cookie';
 import { lightTheme, darkTheme } from '@components/Dark-Theme/Themes';
 
 export const useDarkMode = () => {
   const [theme, setTheme] = useState('light');
   const [themeData, setThemeData] = useState(lightTheme);
   const [mountedComponent, setMountedComponent] = useState(false);
-  const setCookie = (name, value, days) => {
-    const domain = '.realdevsquad.com';
-    const expires = new Date(Date.now() + 24 * days * 60 * 60 * 1000);
-    const cookieStr = `${name}=${value}; expires=${expires}; domain=${domain}; path=/`;
-    document.cookie = cookieStr;
-  };
-  const accessCookie = (cookieName) => {
-    const name = `${cookieName}=`;
-    const allCookieArray = document.cookie.split(';');
-    for (let i = 0; i < allCookieArray.length; i += 1) {
-      const temp = allCookieArray[i].trim();
-      if (temp.indexOf(name) === 0)
-        return temp.substring(name.length, temp.length);
-    }
-    return '';
-  };
   const setMode = (mode) => {
     setCookie('theme', mode, 30);
     setTheme(mode);
@@ -34,7 +19,7 @@ export const useDarkMode = () => {
   };
 
   useEffect(() => {
-    const localTheme = accessCookie('theme');
+    const localTheme = getCookie('theme');
     const themeMode = localTheme === 'light' ? lightTheme : darkTheme;
     if (localTheme) {
       setTheme(localTheme);
